Cache invalidation, better configurable, fragments support.
Cache is updated after configured amount of time, implemented as normal image download, but with if-modified-since HTTP header.
Cache files are also deleted when unused for some time.

You can now disable all types of cache (memory, internal, external).

Witch fragments runOnUIThread doesn't work, updated to handler.

+++ src/com/fedorvlasov/lazylist/FileCache.java
@@ -1,38 +1,92 @@
 package com.fedorvlasov.lazylist;
 
 import java.io.File;
+
+import com.fedorvlasov.lazylist.Utils;
+
 import android.content.Context;
+import android.os.Environment;
 
 public class FileCache {
     
     private File cacheDir;
+    private File tempDir;
     
-    public FileCache(Context context){
+    public FileCache(Context context, boolean allowInternal, boolean allowExternal){
         //Find the dir to save cached images
-        if (android.os.Environment.getExternalStorageState().equals(android.os.Environment.MEDIA_MOUNTED))
-            cacheDir=new File(android.os.Environment.getExternalStorageDirectory(),"LazyList");
-        else
-            cacheDir=context.getCacheDir();
-        if(!cacheDir.exists())
-            cacheDir.mkdirs();
+        if (Environment.getExternalStorageState().equals(Environment.MEDIA_MOUNTED) &&
+        		allowExternal)  {
+        	//external cache
+            cacheDir=new File(Utils.getExternalFilesDir(context),"cache");
+            tempDir = cacheDir;
+        } else if (allowInternal) {
+        	//internal cache
+            cacheDir = context.getCacheDir();
+            tempDir = cacheDir;
+        }
+        else {
+        	//no file cache
+        	cacheDir = null;
+        	tempDir = context.getCacheDir();
+        }	
+        
+        if(cacheDir != null) {
+	        if(!cacheDir.exists())
+	            cacheDir.mkdirs();
+        }
     }
     
     public File getFile(String url){
-        //I identify images by hashcode. Not a perfect solution, good for the demo.
-        String filename=String.valueOf(url.hashCode());
+    	if(cacheDir == null) {
+    		return null;
+    	}
+    	
+        //I identify images by md5 hash of url
+        String filename = Utils.md5(url);
         //Another possible solution (thanks to grantland)
         //String filename = URLEncoder.encode(url);
         File f = new File(cacheDir, filename);
         return f;
-        
+    }
+    
+    public File getTempFile(String url) {
+    	File dir = cacheDir;
+    	if(dir == null) {
+    		dir = tempDir;
+    	}
+    	
+        //I identify images by md5 hash of url
+        String filename = Utils.md5(url) + ".tmp";
+        //Another possible solution (thanks to grantland)
+        //String filename = URLEncoder.encode(url);
+        File f = new File(dir, filename);
+        return f;
     }
     
     public void clear(){
+    	if(cacheDir == null) 
+    		return;
+    	
         File[] files=cacheDir.listFiles();
         if(files==null)
             return;
         for(File f:files)
             f.delete();
     }
+
+	public void clear(long deleteOlderThen) {
+		if(cacheDir == null) {
+			return;
+		}
+		
+		File[] files = cacheDir.listFiles();
+        if(files==null)
+            return;
+        for(File f:files) {
+        	if(f.lastModified() < deleteOlderThen) {
+        		f.delete();        		
+        	}
+        }
+	}
 
 }
